Update Flutter version metadata and web readiness docs
