+++ .metadata
@@ -4,7 +4,7 @@
 # This file should be version controlled and should not be manually edited.
 
 version:
-  revision: "fcf2c11572af6f390246c056bc905eca609533a0"
+  revision: "2c9eb20739dfec95e2c74bd3dfa4601b0a8a36aa"
   channel: "stable"
 
 project_type: app
@@ -18,6 +18,9 @@ migration:
     - platform: android
       create_revision: fcf2c11572af6f390246c056bc905eca609533a0
       base_revision: fcf2c11572af6f390246c056bc905eca609533a0
+    - platform: web
+      create_revision: 2c9eb20739dfec95e2c74bd3dfa4601b0a8a36aa
+      base_revision: 2c9eb20739dfec95e2c74bd3dfa4601b0a8a36aa
 
   # User provided section
